// frontend/src/api/models/analysisRuleCategoryView.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */

export interface AnalysisRuleCategoryView {
  id: string
  primary: string
  detailed: string
  /** @pattern ^#(?:[0-9a-fA-F]{3}|[0-9a-fA-F]{6})$ */
  color: string
  /** @nullable */
  archivedAt?: string | null
}